Move cli from main.rs to cli.rs mod
